DocGen/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Babel==2.12.1
backports.zoneinfo;python_version<"3.9"
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.6
importlib-metadata==6.0.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
pycodestyle==2.10.0
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
zipp==3.15.0
tzdata==2022.7

